move remaining library components to components folder
