+++ src/renderer/library/components/Hardware/HardwareAddress/types.ts
