+++ src/renderer/library/components/Hardware/HardwareStatusBar/types.ts
