+++ src/renderer/library/components/Overlay/index.tsx
