+++ src/renderer/library/components/QRCode/Display.tsx
