+++ src/renderer/library/components/QRCode/DisplayPayload.tsx
