+++ src/renderer/library/components/QRCode/Html5QrCodePlugin.tsx
