+++ src/renderer/library/components/QRCode/Wrappers.ts
