+++ src/renderer/library/components/QRCode/constants.ts
