+++ src/renderer/library/components/QRCode/qrcode.ts
