+++ src/renderer/library/components/QRCode/types.ts
