+++ src/renderer/library/components/QRCode/util.ts
