+++ src/renderer/library/components/Switch/index.tsx
