+++ src/renderer/library/components/index.ts
@@ -8,7 +8,11 @@ export * from './Footer';
 export * from './Header';
 export * from './Identicon';
 export * from './Menu';
+export * from './Overlay';
 export * from './ShiftingMeter';
 export * from './SideNav';
 export * from './SortControls';
 export * from './Stats';
+export * from './Switch';
+export * from './Tooltip';
+export * from './Tx';
